requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.0.8
django-extensions==3.0.5
django-filter==2.3.0
django-js-asset==1.2.2
django-memoize==2.3.1
django-model-utils==4.0.0
django-mptt==0.11.0
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
filelock==3.0.12
flake8==3.8.3
identify==1.5.9
idna==2.10
importlib-metadata==1.7.0
isort==5.2.0
mccabe==0.6.1
nodeenv==1.5.0
pathspec==0.8.0
phonenumbers==8.12.9
Pillow==7.2.0
pre-commit==2.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
virtualenv==20.2.0
zipp==3.1.0
